desktop icon
Added new desktop icon and change function of auto startup script (that runs on
startup and also runs from the desktop icon).

The new startup script will just activate the RDE Chromium window if it is
running, or start it from scratch if it is not.

In addition, startup script is now a link, to make it easier to replace startup
script when running auto boot programs (like the simon demo)